rename: view 내부 패키지 제거
