+++ src/main/java/controller/LadderGameController.java
@@ -2,8 +2,8 @@ package controller;
 
 import domain.*;
 import util.BooleanGenerator;
-import view.input.InputView;
-import view.output.OutputView;
+import view.InputView;
+import view.OutputView;
 
 public class LadderGameController {
