+++ src/main/java/view/InputView.java
@@ -1,4 +1,4 @@
-package view.input;
+package view;
 
 import java.util.Scanner;
